Squeezing out a bit of code reuse.
Switching editing to a separate screen, because having the item editor render
in a sheet over the top of the massive scrolling list is excruciatingly slow
(on macOS, not so bad on iOS). Also made the list less massively long by going
back to showing just recently completed items and hiding the full history list
behind a toggle button.